Move currency library to lerna module
