+++ packages/currency/src/index.js
